% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_hist.R
\name{arrow_hist}
\alias{arrow_hist}
\title{Generate histograms of lambda distributions with red line indicating the
lambda statistic for the discovery analysis transprobe SNPs}
\usage{
arrow_hist(discsnps_p, lambda, resamp_n = 1000, out_dir = ".")
}
\arguments{
\item{discsnps_p}{Vector of discovery analysis transprobe SNPs}

\item{lambda}{Dataframe with columns of mean and median lambdas, per resamp
Colnames as "median_lambda", and "mean_lambda"}

\item{resamp_n}{Integer for number of resampling performed}

\item{out_dir}{Output directory}
}
\value{
1) Lambda for the discovery analysis transprobe SNPs AND
 2) A median and mean lambda distribution histogram with red lines
 for the lambda statistic from the discovery analysis transprobe SNPs,
 written to the current working directory.
}
\description{
Generate histograms of lambda distributions with red line indicating the
lambda statistic for the discovery analysis transprobe SNPs
}
\author{
Chloe X Yap
}

