% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_circleplot.R
\name{circle_plot}
\alias{circle_plot}
\title{Produce a circle plot showing trans-eQTL SNP/probe associations.}
\usage{
circle_plot(df, gws_snps = NA, id_order = qq_out[["id_ordered.row"]],
  hg_chr = seq_name, hg_len = seq_length, gen_name = "hg19")
}
\arguments{
\item{df}{Dataframe of 
Column names must include:
  1) "rsid"
  2) "snpchr"
  3) "snpbp"
  4) "probe"
  5) "probechr"
  6) "probebp"
  7) "qqid"
  8) "p_eqtl" N.B. this is the p-value for the trans-eQTL association
  9) "p_snp"}

\item{gws_snps}{Dataframe with GWS SNPs
Column names must include:
  1) "rsid"
  2) "p_snp"
  3) "snpchr"
  4) "snpbp"
If input is NA, then GWS SNPs will not be shown on the circle plot.}

\item{id_order}{If qq_plot() function was run, input out[["id-ordered.row"]]
Vector of row numbers corresponding to the original "df" order, but ordered 
by descending -log10 p-value of the sentinel trans-gene SNP. 
Eg. c(3, 1, 2) if the trans-eQTL association represents that the trans-eQTL
association in the 3rd row of df had the strongest sentinel SNP (-log10(p))}

\item{hg_chr}{Use for GRanges object seqname
Extract using code:
hg19chr <- fetchExtendedChromInfoFromUCSC("hg19")
seq_name <- hg19chr[1:22,"NCBI_seqlevel"]}

\item{gen_name}{Use for GRanges genome name. Default = "hg19"}

\item{hg_chr}{Use for GRanges object seqlength
Extract using code:
hg19chr <- fetchExtendedChromInfoFromUCSC("hg19")
seq_length <- hg19chr[1:22, "UCSC_seqlength"]}

\item{legend}{Whether to put legend on the circle plot
One from: c(FALSE, TRUE, "both")}
}
\value{
Circle plot showing for each trans-eQTL association the positions of 
 the SNP (red point) and probe (green line). The links connecting these
 associations are colour-coded to match the transprobe regions on the
 compartmentalised QQ plot. Their size is proportional to the strength of the 
 eQTL association. If an additional dataframe of GWS SNPs and their  
 positions are input into the gws_snps argument, they will be plotted as 
 a Manhattan plot on a track.
}
\description{
Produce a circle plot showing trans-eQTL SNP/probe associations.
}
\author{
Chloe X Yap
}

