% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_snps.R
\name{lambda_snps}
\alias{lambda_snps}
\title{Calculate lambda statistic (median and mean) for a vector of SNPs}
\usage{
lambda_snps(snps)
}
\arguments{
\item{snps}{Dataframe, with a column titled "p_snp" with SNP p-values}
}
\value{
Vector of median and mean lambdas for that vector of p-values
}
\description{
Calculate lambda statistic (median and mean) for a vector of SNPs
}
\author{
Chloe X Yap
}

