% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_pQQ2.fun.R
\name{points_pQQ2.fun}
\alias{points_pQQ2.fun}
\title{QQ plot points to overlay onto the "base" plot from pQQ2.fun.lite()}
\usage{
points_pQQ2.fun(pvals, resamp = TRUE, colour = "orange", subset.qq = NA,
  nlabs.2 = NA, pch_point = ".", ...)
}
\arguments{
\item{pvals}{Vector of p-values}

\item{resamp}{TRUE or FALSE; 
If resamp = TRUE, get smaller points
If resamp = FALSE, this function is meant to be called within the 
qqplot_comp() function. This will make points bigger, and will assign
GWAS SNPs corresponding to different transprobe regions different colours}

\item{colour}{Colour for points. Default is "orange", though if
resamp = TRUE, different colours will be assigned per transprobe region.}

\item{subset.qq}{Only relevant when resamp = FALSE. Labels GWAS SNP with the 
lowest p-value in each transprobe region with its rsid.}

\item{nlabs.2}{Only relevant when resamp = FALSE. Determines how many rsid 
labels to have (see subset.qq)}

\item{pch_point}{Point size. Default is '.', - tiny point for resamp = TRUE}
}
\value{
QQ plot points, overlaid on "base" QQ plot from pQQ2.fun.lite()
}
\description{
N.B. This code is adapted from Haplin::pQQ
\url{http://www.inside-r.org/packages/cran/Haplin/docs/pQQ}
It has only been slightly adapted to suit the purposes of this pipeline.
It is placed within this package for convenience of access.
}
\details{
This function is also used within the following functions:
\code{\link{qqplot_comp}}
\code{\link{sep_qq}}
}
\author{
adapted from Haplin::pQQ
}

