% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQQ2.fun.lite.R
\name{pQQ2.fun.lite}
\alias{pQQ2.fun.lite}
\title{QQ plot, also acts as a base on which to overlay points_pQQ2.fun()}
\usage{
pQQ2.fun.lite(pvals, colour = "gray20", nlabs = 6, conf = 0.95, lim,
  mark = 0.05, limx = c(0, log10(length(pvals)) + 0.5), ...)
}
\arguments{
\item{pvals}{Vector of p-values}

\item{colour}{Colour for points}

\item{lim}{y-axis limits. If overlaying points, default may not be optimal}

\item{limx}{x-axis limits. If overlaying points, default may not be optimal
\url{http://www.inside-r.org/packages/cran/Haplin/docs/pQQ}}
}
\value{
QQ plot
}
\description{
N.B. This code is adapted from Haplin::pQQ
\url{http://www.inside-r.org/packages/cran/Haplin/docs/pQQ}
It has only been slightly adapted to suit the purposes of this pipeline.
It is placed within this package for convenience of access.
}
\details{
This function is also used within the following functions:
\code{\link{qqplot_comp}}
\code{\link{sep_qq}}
}
\author{
adapted from Haplin::pQQ
}

