% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_comp.R
\name{qqplot_comp}
\alias{qqplot_comp}
\title{Plot QQ-plot with GWAS SNP p-values compartmentalised by qqid.}
\usage{
qqplot_comp(snps, name = "out", out_dir = ".")
}
\arguments{
\item{snps}{Dataframe of trans-gene SNPs;
Require columns with names:
1) "p_snp" - for p-value       (colClass: numeric)
2) "qqid" - for identifier (colClass: character)
3) "rsid" - for SNP ID     (colCLass: character)}

\item{name}{Character: phenotype name.}

\item{Directory}{for output. Default is current working directory}
}
\value{
QQ plot, with SNPs compartmentalised by trans-gene region.
   The colours (from red to purple on the colour wheel) are ordered 
   by ascending -log10 p-value of the sentinel SNP in the region. 
   The sentinel SNPs are also labelled by their rsid on the plot.
   A list is also returned as output containing:
     1) Vector of the colour IDs used.
     2) Dataframe of the qqids, ordered by -log10(p-value) of the 
        sentinel trans-gene SNP.
     3) Vector of row indices, ordered by -log10(p-value) of the 
        sentinel trans-gene SNP. Used to construct circle_plot.
}
\description{
BEFORE using this function, pQQ2.fun.lite() and points_pQQ2.fun()
need to be loaded. These functions are available in this package:
\code{\link{pQQ2.fun.lite}}
\code{\link{points_pQQ2.fun}}
In turn, these QQ plot function are slightly-modified versions 
of Haplin::pQQ
\url{http://www.inside-r.org/packages/cran/Haplin/docs/pQQ}
}
\author{
Chloe X Yap
}

