% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_qq.R
\name{sep_qq}
\alias{sep_qq}
\title{Plot an individual QQ-plot for each trans-gene region, and produce array.}
\usage{
sep_qq(snps, id_ordered = qq_out[["id_ordered"]]$qqid,
  colouryay = qq_out[["colouryay"]])
}
\arguments{
\item{snps}{Dataframe of trans-gene SNPs;
Require columns with names:
1) "p_snp" - for p-value   (colClass: numeric)
2) "qqid" - for identifier (colClass: character)
3) "rsid" - for SNP ID     (colCLass: character)}

\item{id_ordered}{Vector of qqids. Required to match transprobe SNP p-values
to the probe. Default: "qq_out[["id_ordered"]]$qqid".}

\item{colouryay}{Vector of colour IDs corresponding to each 
transprobe region. Length must match length(id_ordered).
Default: use directly from qq_plot(resamp = F) output so colours match 
ie. "qq_out[["colouryay"]]"}
}
\value{
Arrays of QQ plots. Colours used will match those representing each
 trans-gene region from the qq_plot function. QQ plots into automatically 
 be grouped into pages of 6 or 4 trans-eQTL associations, depending on 
 if the total number of associations is better divided by 6 or 4.
}
\description{
BEFORE using this function, pQQ2.fun.lite() and points_pQQ2.fun()
need to be loaded. These functions are available in this package:
\code{\link{pQQ2.fun.lite}}
\code{\link{points_pQQ2.fun}}
In turn, these QQ plot function are slightly-modified versions 
of Haplin::pQQ
\url{http://www.inside-r.org/packages/cran/Haplin/docs/pQQ}
}
\author{
Chloe X Yap
}

