% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_cister_eqtls.R
\name{trans_cister_eqtls}
\alias{trans_cister_eqtls}
\title{Find trans-eQTLs and cister-eQTLs}
\usage{
trans_cister_eqtls(eqtlfile, eqtl_cols = c(2, 15, 1, 10, 7, 9),
  query_snps = "all", sig = 1e-06, resamp = FALSE, write = TRUE,
  out_dir = ".", name = "out", keep_allsig = FALSE)
}
\arguments{
\item{eqtlfile}{Dataframe of eqtls. Cols and classes described in eqtl_cols.}

\item{eqtl_cols}{Vector of eqtlfile column indices in order: 
1) SNP rsid     (colClass: character)
2) eQTL p-value (colClass: numeric)
3) SNP chr      (colClass: integer)
4) probe chr    (colClass: integer)
5) probe ID     (colClass: character)
6) gene name    (colClass: character).}

\item{query_snps}{Vector of query SNPs (class: character); 
if already extracted, input "all" (default).}

\item{sig}{Numeric: eQTL p-value significance threshold. Default = 1e-6}

\item{resamp}{TRUE or FALSE;
FALSE value creates tables of sig eQTLs, trans-eQTLs, cister-eQTLs and 
trans-eQTL IDs in out_dir folder. Default is TRUE;
TRUE does not create any tables; Use FALSE with resampling analysis.}

\item{write}{TRUE or FALSE;
Variable only required if resamp = FALSE;
TRUE value writes all sig eQTLs, trans-eQTLs, cister-eQTLs and 
trans-eQTL IDs to the specified out_dir. 
FALSE value means that out_dir and name are not required variables.}

\item{out_dir}{Character: output directory. Default = "./"}

\item{name}{Character: Used for naming output files 
Eg. phenotype. Default = "out"}

\item{keep_allsig}{TRUE or FALSE;
TRUE value stores all eQTLs with p-value < "sig" in the output list.
FALSE recommended for resampling analysis; Default is FALSE.}
}
\value{
If resamp = FALSE; return list with entries for
   all sig eQTLs, trans-eQTLs, cister-eQTLs, and trans-eQTL IDs 
   (N.B. trans-eQTL IDs only show trans-eQTLs with autosomal probes);
   If resamp = TRUE, return list with trans-eQTLs associations, 
   number of cis-eQTLs,
   number of trans-eQTL SNPs.
   In addition, keep_allsig = TRUE returns a data frame of all significant 
   eQTLs as a list entry.
}
\description{
Find trans-eQTLs and cister-eQTLs
}
\examples{
trans_cister_eqtls(eqtl_df, 
                   eqtl_cols = c(1:6), 
                   query_snps = "all", 
                   sig = 1e-6, 
                   out_dir = "Output",                  
                   name = "out",
                   resamp = FALSE,
                   keep_allsig = FALSE)

}
\author{
Chloe X Yap
}

